requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.52.0
PyPDF2==3.0.1
langchain==0.1.12
faiss-cpu==1.7.4
